Refactor sync classes
