+++ src/peergos/server/sync/PeergosSyncFS.java
@@ -96,7 +96,7 @@ public class PeergosSyncFS implements SyncFilesystem {
     }
 
     @Override
-    public DirectorySync.Blake3state hashFile(Path p) {
+    public Blake3state hashFile(Path p) {
         byte[] buf = new byte[4 * 1024];
         Blake3 state = Blake3.initHash();
 
@@ -110,7 +110,7 @@ public class PeergosSyncFS implements SyncFilesystem {
         }
 
         byte[] hash = state.doFinalize(32);
-        return new DirectorySync.Blake3state(hash);
+        return new Blake3state(hash);
     }
 
     @Override
